refactor(order): 將 Entity/Repository 子 package 扁平化至 order root
與 account/product 一致改用「扁平 package + 類名後綴」風格，
移除違反 Java 命名慣例的大寫 Entity/、Repository/ 子 package。
DTO/ 維持不動。
