+++ src/test/java/com/ibm/demo/OptimisticLockingIntegrationTest.java
@@ -16,8 +16,8 @@ import com.ibm.demo.account.AccountRepository;
 import com.ibm.demo.enums.AccountStatus;
 import com.ibm.demo.enums.OrderStatus;
 import com.ibm.demo.enums.ProductStatus;
-import com.ibm.demo.order.Entity.OrderInfo;
-import com.ibm.demo.order.Repository.OrderInfoRepository;
+import com.ibm.demo.order.OrderInfo;
+import com.ibm.demo.order.OrderInfoRepository;
 import com.ibm.demo.product.Product;
 import com.ibm.demo.product.ProductRepository;
