+++ src/test/java/com/ibm/demo/order/OrderReadEndpointOsivIntegrationTest.java
@@ -24,9 +24,6 @@ import com.ibm.demo.enums.OrderStatus;
 import com.ibm.demo.enums.ProductStatus;
 import com.ibm.demo.order.DTO.GetOrderDetailResponse;
 import com.ibm.demo.order.DTO.GetOrderListResponse;
-import com.ibm.demo.order.Entity.OrderDetail;
-import com.ibm.demo.order.Entity.OrderInfo;
-import com.ibm.demo.order.Repository.OrderInfoRepository;
 import com.ibm.demo.product.Product;
 import com.ibm.demo.product.ProductClient;
 import com.ibm.demo.product.ProductRepository;
