+++ src/test/java/com/ibm/demo/order/OrderServiceTest.java
@@ -36,8 +36,6 @@ import com.ibm.demo.order.DTO.OrderDeletionPlan;
 import com.ibm.demo.order.DTO.OrderView;
 import com.ibm.demo.order.DTO.UpdateOrderDetailRequest;
 import com.ibm.demo.order.DTO.UpdateOrderRequest;
-import com.ibm.demo.order.Entity.OrderInfo;
-import com.ibm.demo.order.Repository.OrderInfoRepository;
 import com.ibm.demo.product.DTO.internal.AdjustStockRequest;
 import com.ibm.demo.product.DTO.internal.OrderItemRequest;
 import com.ibm.demo.product.ProductClient;
